Move resources to their own package
