+++ src/server/Resources/Customer.java
@@ -2,7 +2,7 @@
 // Kevin T. Manley
 // CSE 593
 // -------------------------------
-package server.ResImpl;
+package server.Resources;
 
 import java.util.Enumeration;
