+++ src/server/Resources/RMItem.java
@@ -2,7 +2,7 @@
 // adapted from Kevin T. Manley
 // CSE 593
 // -------------------------------
-package server.ResImpl;
+package server.Resources;
 
 import java.io.Serializable;
